feat(port): remove method open from port interface
Since the open method depends on "serial.Config", the port interface always
contains a dependency to the serial package. To prevent this, the method
is removed. This change has no side effects.
The usage has not changed and can be tested with the example and the existing
unit test.
For this purpose virtual interfaces can be created with socat.
"socat -d -d pty,link=/dev/ttys009,raw,echo=0 pty,link=/dev/ttys010,raw,echo=0"